Revert "test parallelization"
This reverts commit 8a33dd1bd109af90bdc85fbde8dfc4d4c5dc980e.
