+++ src/EphemeralMongo.Tests/MongoRunnerTests.cs
@@ -1,3 +1,10 @@
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.InteropServices;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using EphemeralMongo.Download;
+using MongoDB.Driver;
+
 namespace EphemeralMongo.Tests;
 
 public class MongoRunnerTests(ITestOutputHelper testOutputHelper, ITestContextAccessor testContextAccessor)
@@ -7,8 +14,8 @@ public class MongoRunnerTests(ITestOutputHelper testOutputHelper, ITestContextAc
     {
         var options = new MongoRunnerOptions
         {
-            StandardOutputLogger = line => MongoTrace.LogToOutput(line, testOutputHelper),
-            StandardErrorLogger = line => MongoTrace.LogToOutput(line, testOutputHelper),
+            StandardOutputLogger = this.MongoMessageLogger,
+            StandardErrorLogger = this.MongoMessageLogger,
             BinaryDirectory = Guid.NewGuid().ToString(),
             AdditionalArguments = ["--quiet"],
         };
@@ -40,8 +47,8 @@ public class MongoRunnerTests(ITestOutputHelper testOutputHelper, ITestContextAc
         {
             var options = new MongoRunnerOptions
             {
-                StandardOutputLogger = line => MongoTrace.LogToOutput(line, testOutputHelper),
-                StandardErrorLogger = line => MongoTrace.LogToOutput(line, testOutputHelper),
+                StandardOutputLogger = this.MongoMessageLogger,
+                StandardErrorLogger = this.MongoMessageLogger,
                 RootDataDirectoryPath = rootDataDirectoryPath,
                 AdditionalArguments = ["--quiet"],
             };
@@ -88,6 +95,159 @@ public class MongoRunnerTests(ITestOutputHelper testOutputHelper, ITestContextAc
             catch (DirectoryNotFoundException)
             {
             }
+        }
+    }
+
+    [Theory]
+    [InlineData(false, MongoVersion.V6, MongoEdition.Community)]
+    [InlineData(false, MongoVersion.V7, MongoEdition.Community)]
+    [InlineData(false, MongoVersion.V8, MongoEdition.Community)]
+    [InlineData(false, MongoVersion.V6, MongoEdition.Enterprise)]
+    [InlineData(false, MongoVersion.V7, MongoEdition.Enterprise)]
+    [InlineData(false, MongoVersion.V8, MongoEdition.Enterprise)]
+    [InlineData(true, MongoVersion.V6, MongoEdition.Community)]
+    [InlineData(true, MongoVersion.V7, MongoEdition.Community)]
+    [InlineData(true, MongoVersion.V8, MongoEdition.Community)]
+    [InlineData(true, MongoVersion.V6, MongoEdition.Enterprise)]
+    [InlineData(true, MongoVersion.V7, MongoEdition.Enterprise)]
+    [InlineData(true, MongoVersion.V8, MongoEdition.Enterprise)]
+    public async Task Import_Export_Works(bool replset, MongoVersion version, MongoEdition edition)
+    {
+        if (version is MongoVersion.V6 or MongoVersion.V7 && edition == MongoEdition.Enterprise && RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+        {
+            var target = DownloadTargetHelper.Instance.GetOriginalLinuxTarget();
+            if (target == "ubuntu2404")
+            {
+                Assert.Skip("Enterprise MongoDB 6 and 7 are not supported on Ubuntu 24.04");
+            }
+        }
+
+        const string databaseName = "default";
+        const string collectionName = "people";
+
+        var options = new MongoRunnerOptions
+        {
+            Version = version,
+            Edition = edition,
+            UseSingleNodeReplicaSet = replset,
+            StandardOutputLogger = this.MongoMessageLogger,
+            StandardErrorLogger = this.MongoMessageLogger,
+            AdditionalArguments = edition == MongoEdition.Enterprise
+                ? ["--quiet", "--storageEngine", "inMemory"]
+                : ["--quiet"],
+        };
+
+        using (var runner = await MongoRunner.RunAsync(options, testContextAccessor.Current.CancellationToken))
+        {
+            if (replset)
+            {
+                Assert.Contains("replicaSet", runner.ConnectionString);
+            }
+            else
+            {
+                Assert.DoesNotContain("replicaSet", runner.ConnectionString);
+            }
+        }
+
+        var originalPerson = new Person("john", "John Doe");
+        var exportedFilePath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
+
+        try
+        {
+            using (var runner1 = await MongoRunner.RunAsync(options, testContextAccessor.Current.CancellationToken))
+            {
+                var database = new MongoClient(runner1.ConnectionString).GetDatabase(databaseName);
+
+                // Verify that the collection is empty
+                var personBeforeImport = database.GetCollection<Person>(collectionName).Find(FilterDefinition<Person>.Empty).FirstOrDefault(testContextAccessor.Current.CancellationToken);
+                Assert.Null(personBeforeImport);
+
+                // Add a document
+                await database.GetCollection<Person>(collectionName).InsertOneAsync(new Person(originalPerson.Id, originalPerson.Name), options: null, cancellationToken: testContextAccessor.Current.CancellationToken);
+                await runner1.ExportAsync(databaseName, collectionName, exportedFilePath, ["--jsonArray"], testContextAccessor.Current.CancellationToken);
+
+                // Verify that the document was inserted successfully
+                var personAfterImport = database.GetCollection<Person>(collectionName).Find(FilterDefinition<Person>.Empty).FirstOrDefault(testContextAccessor.Current.CancellationToken);
+                Assert.Equal(originalPerson, personAfterImport);
+            }
+
+            IMongoRunner runner2;
+            using (runner2 = await MongoRunner.RunAsync(options, testContextAccessor.Current.CancellationToken))
+            {
+                var database = new MongoClient(runner2.ConnectionString).GetDatabase(databaseName);
+
+                // Verify that the collection is empty
+                var personBeforeImport = database.GetCollection<Person>(collectionName).Find(FilterDefinition<Person>.Empty).FirstOrDefault(testContextAccessor.Current.CancellationToken);
+                Assert.Null(personBeforeImport);
+
+                // Import the exported collection
+                await runner2.ImportAsync(databaseName, collectionName, exportedFilePath, ["--jsonArray"], cancellationToken: testContextAccessor.Current.CancellationToken);
+
+                // Verify that the document was imported successfully
+                var personAfterImport = database.GetCollection<Person>(collectionName).Find(FilterDefinition<Person>.Empty).FirstOrDefault(testContextAccessor.Current.CancellationToken);
+                Assert.Equal(originalPerson, personAfterImport);
+            }
+
+            // Disposing twice does nothing
+            runner2.Dispose();
+
+            // Can't use import or export if already disposed
+            await Assert.ThrowsAsync<ObjectDisposedException>(() => runner2.ExportAsync("whatever", "whatever", "whatever.json", cancellationToken: testContextAccessor.Current.CancellationToken));
+            await Assert.ThrowsAsync<ObjectDisposedException>(() => runner2.ImportAsync("whatever", "whatever", "whatever.json", cancellationToken: testContextAccessor.Current.CancellationToken));
+        }
+        finally
+        {
+            File.Delete(exportedFilePath);
+        }
+    }
+
+    private void MongoMessageLogger(string message)
+    {
+        try
+        {
+            var trace = JsonSerializer.Deserialize<MongoTrace>(message);
+
+            if (trace != null && !string.IsNullOrEmpty(trace.Message))
+            {
+                // https://www.mongodb.com/docs/manual/reference/log-messages/#std-label-log-severity-levels
+                var logLevel = trace.Severity switch
+                {
+                    "F" => "CTR",
+                    "E" => "ERR",
+                    "W" => "WRN",
+                    _ => "INF",
+                };
+
+                const int longestComponentNameLength = 8;
+                testOutputHelper.WriteLine("{0} {1} {2}", logLevel, trace.Component.PadRight(longestComponentNameLength), trace.Message);
+                return;
+            }
+        }
+        catch (JsonException)
+        {
+        }
+
+        testOutputHelper.WriteLine(message);
+    }
+
+    private sealed class MongoTrace
+    {
+        [JsonPropertyName("s")]
+        public string Severity { get; set; } = string.Empty;
+
+        [JsonPropertyName("c")]
+        public string Component { get; set; } = string.Empty;
+
+        [JsonPropertyName("msg")]
+        public string Message { get; set; } = string.Empty;
+    }
+
+    private sealed record Person(string Id, string Name)
+    {
+        [SuppressMessage("ReSharper", "UnusedMember.Local", Justification = "Used by MongoDB deserialization")]
+        public Person()
+            : this(string.Empty, string.Empty)
+        {
         }
     }
 }
