+++ src/EphemeralMongo.Tests/xunit.runner.json
@@ -1,3 +1,3 @@
 {
-  "$schema": "https://xunit.net/schema/current/xunit.runner.schema.json"
+    "$schema": "https://xunit.net/schema/current/xunit.runner.schema.json"
 }
